Renamed `RomanNumeral` struct to `Numerus`.
Changed `Numerus` from a tuple struct to a struct with one private member, `vis` ("strength, value") that holds and hides the underlying `u16`.

Implemented the `Default` trait for `Numerus` with a default value of 1.  Updated the `Display` trait, making values 0 and >4000 invalid and adding display values "I" through "X".

Implemented the `From` trait for `Numerus` to `u16` conversions and implemented the `TryFrom` trait for `u16` to `Numerus` conversions, failing for 0 and >4000.  Created error struct `Nimis` ("too much") as the `TryFrom` error type.