Implemented the `FromStr` trait for `Numerus`.
The first-pass algorithm scans characters from the start of the string and accumulates the integer value based on the current character and the previous character to account for "IV", "IX", et cetera.  The parse only accepts capital letters "IVXLCDM" and fails if any other letters are present.  It also fails for malformed Roman numerals like "VX".