Implement a custom `Debug` for `Orthographia`.
Instead of the default debug representation, which prints the internal
`String` representation, including long vowels as Unicode characters
with macrons, simply print out the "ASCII" representation of the
internal `String`, with surrounding double quotes.  Thus the debug
representation of _māter_ is `"ma'ter"`.

Plus a couple of small formatting changes.