Refactor the `try_from_ascii()` and `try_from_canonical()` methods.
Simplify the `try_from_ascii()` and `try_from_canonical()` methods of
`Orthographia` by creating chainable failable iterators for scanning the
incoming character string and doing the requisite checks and
transformations.  These iterators use an item type of `Result<char,
Irritus>` and these chains rely on that `collect()` can yield a final
result of `Result<String, Irritus>` from the iterator chain.

The iterator structs are `AsciiChars` and `CanonicalChars` for checking
the allowed character set, `InitialCaps` and `SoloHyphen` for checking
permitted capitalization and denying repeated hyphens,
`LongVowelMacrons` and `LongVowelTicks` for transforming long vowel
sequences into precombined macron vowel characters, and `NotEmpty` for
checking that the source string contains one or more characters.

The `CharFilter` iterator adapts a `Chars` iterator so that it returns a
`Result` instead of a `char`.

The `CharFilters` trait adds friendly methods to the iterators for easy
chaining, and the `ToCharFilter` trait adds a `char_filter()` method to
`&str`.