Added traits `Vocabulum` and `Forma`.
The `Vocabulum` represents a word with a part of speach and one or more
forms ("formae").  A `Vocabulum` also has a "lemma", which is the
particular form used as the headword in a dictionary or other reference.

The `Forma` trait represents one of the forms of a `Vocabulum`.  It has
an `Orthographia` ("spelling") and a link to the `Vocabulum` it belongs
to.

Different parts of speach have very different numbers of forms: nouns
have 10 (or 12 with the vocative), adjectives 20 or 30, verbs ~150,
conjunctions only one.  Thus it makes sense for each part of speach to
have its own concrete representation and use a dyn trait for those
places where words and forms need a uniform interface.