Added structs `Nomen` and `FormaNominis`.
`Nomen` represents a noun word and implements the `Vocabulum` trait.  It
contains a list of `FormaNominis` children that implement the `Forma`
trait.

Added enums for noun attributes `Declinatio`, `Genus`, `Casus` and
`Numerus`.

Added `Nomen` creational method `try_from_ascii_stem()` and implemented
the creation of regular first declention nouns to start.

Tweaked the implementation of `FormaConiunctionis` to use a `*const
Coniunctio` to point to its parent instead of a `*mut Coniunctio`.