Allow vertical line (`|`) chars in `Orthographia`.
Modify the internal string representation of `Orthographia` to allow the
use of `|` to separate word stems from endings, as in `vacc|a`.  In
addition, also allow a leading `|` to mark a stand-alone word ending
like `|us` and a trailing `|` to mark a word stem like `nas|`.

Modified fallible char iterators `AsciiChars` and `CanonicalChars` to
accept the `|` char.  Added fallible iterator `SoloPipes` to disallow
two `|` chars in a row.

For output formats, added `NoStemEndingSeparators` to strip `|` chars
falling between two letters.  Added `EndingHyphens` and `StemHyphens` to
transform `|` to `-` for stand-alone word endings and stems.

Revised the doc comments for `Orthographia` along with its `Debug` and
`Display` impl docs.