Added struct `Locutio` ("phrase").
A `Locutio` holds one or more _Orthographiae_ that are written as
separate "words" but make up a single word form, like "amatus es"
(passivum indicativus perfectum) or "viginti duo" (XXII).

Added two constructors: `from_word()` for the common case where a
`Locutio` is a single `Orthographia` and `new()` for the general case.