Simplify the implementation of `numerum_fac_e_filo()`.
No need to validate the length of the input string since by the nature
of the `segmentum_numeri` tabulae, processing of the input string is
bounded and will at most process 15 characters and end.

Also, due to this bounding of input, the `summa` will never exceed
`NUMERUS_MAX.vis`, so the only possible error returned is
`error_datis_vitiosis`.  Simplify the error test cases by only checking
for this error.