Add `numero_loca_numerum_cardinalem()` function.
Began work on a function to build the cardinal number phrase for a given
number and implemented cardinal numbers up to 37 ("triginta septem").

Added `enum genus` with cases for no gender, masculinum, femininum and
neutrum.  Added notes on Latin word genders in the docs.

Added struct `linea_mutabilis` (mutable string) and supporting functions
for building cardinal number strings.

Added a static `xstrdup()` fail-fast wrapper of `strdup()` and used it
in the existing string functions for `numerus`.